+++ src/app/cart/components/cart-item-component.ts/cart-item.component.spec.ts
@@ -1,16 +1,16 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { FirstComponent } from './first.component';
+import { CartItemComponent } from './cart-item.component';
 
-describe('FirstComponent', () => {
-  let component: FirstComponent;
-  let fixture: ComponentFixture<FirstComponent>;
+describe('CartItemComponent', () => {
+  let component: CartItemComponent;
+  let fixture: ComponentFixture<CartItemComponent>;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [FirstComponent]
+      declarations: [CartItemComponent],
     });
-    fixture = TestBed.createComponent(FirstComponent);
+    fixture = TestBed.createComponent(CartItemComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });