+++ src/app/products/components/product-component/product.component.ts
@@ -1,20 +1,17 @@
-import { Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { ProductModel } from '../../interfaces/product-model';
-import { CartService } from '../../../cart/services/cart.service';
-import { take } from 'rxjs';
+
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ProductComponent {
   @Input() product!: ProductModel;
-
-  constructor(private readonly cartService: CartService) {}
+  @Output() addToCart = new EventEmitter<ProductModel>();
 
   public onAddToCart(): void {
-    this.cartService.cart$.pipe(take(1)).subscribe((cart) => {
-      this.cartService.cart$.next([...cart, this.product.name]);
-    });
+    this.addToCart.emit(this.product);
   }
 }