+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,27 @@
+import { Directive, ElementRef, HostListener, Renderer2 } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]',
+})
+export class HighlightDirective {
+  color = 'yellow';
+  constructor(
+    private readonly elementRef: ElementRef<HTMLElement>,
+    private readonly renderer: Renderer2,
+    ) {
+  }
+
+  @HostListener('mouseenter') onMouseEnter() {
+    this.highlight(this.color);
+  }
+
+  @HostListener('mouseleave') onMouseLeave() {
+    this.highlight('');
+  }
+
+  private highlight(color: string) {
+    this.renderer.setStyle(this.elementRef.nativeElement, 'background-color', color);
+  }
+}
+
+