+++ changelog.md
@@ -1,5 +1,15 @@
-08.08.2023 : 
+08.08.2023 :
 - added FirstComponent;
 - added product and cart components;
 - added services for product and cart components;
 - clicking the 'buy' button moves product to cart;
+
+21.08.2023 :
+- moved components to their own moduler;
+- if product is not available the button 'buy' is disabled;
+- added totalCost and totalQuantity to cart;
+- added quantity of several type product in cart;
+- added buttons to add, substract or remove items from cart;
+- moved components to OnPush strategy;
+- set the title from the component;
+- added Highlight directive;